aparently forge config system can't handle long, so use int or double as appropriate

+++ src/main/java/com/mordenkainen/equivalentenergistics/core/config/EqEConfig.java
@@ -60,21 +60,21 @@ public final class EqEConfig {
 
         @Config.RangeInt(min = 0, max = Integer.MAX_VALUE)
         @Config.Comment("The amount of EMC the basic EMC Condenser can convert per tick. This value is scaled by x10 for each tier of Condenser.")
-        public long emcPerTick = 8192;
+        public int emcPerTick = 8192;
 
     }
 
     public static class EMCCellCaps {
 
-        public long creativeCell = 16384000000L;
-        public long tier1Cell = 1000000L;
-        public long tier2Cell = 4000000L;
-        public long tier3Cell = 16000000L;
-        public long tier4Cell = 64000000L;
-        public long tier5Cell = 256000000L;
-        public long tier6Cell = 1024000000L;
-        public long tier7Cell = 4096000000L;
-        public long tier8Cell = 16384000000L;
+        public double creativeCell = 16384000000L;
+        public double tier1Cell = 1000000L;
+        public double tier2Cell = 4000000L;
+        public double tier3Cell = 16000000L;
+        public double tier4Cell = 64000000L;
+        public double tier5Cell = 256000000L;
+        public double tier6Cell = 1024000000L;
+        public double tier7Cell = 4096000000L;
+        public double tier8Cell = 16384000000L;
 
     }
